+++ crates/pm/src/service/mod.rs
@@ -1,4 +1,5 @@
 pub mod binary;
+pub mod clean;
 pub mod config;
 pub mod dependency_graph;
 pub mod execute;
